* Notify other players in the room when a new player joins the room. * Fix cast exception error in KatanaClient

+++ trunk/src/server/KatanaClient.java
@@ -32,7 +32,7 @@ public class KatanaClient implements Runnable
     }
     
     public void setId(int id) { this.id = id; }
-    public long getId() { return id; }
+    public int getId() { return id; }
     
     public void remove()
     {
